refactor(core): share skill loader
